Completed Payroll Variables Setting Page

+++ Domain/Deduction.cs
@@ -6,6 +6,6 @@ namespace Domain
     {
         public Guid Id { get; set; }
         public string Name { get; set; }
-        public int Percentage { get; set; }
+        public float Percentage { get; set; }
     }
 }
